Tutorial 1
Added new background.

Added the scenes to the build settings.
Added in the functions to load the tutorial scenes.
Changed the scene after the menu to be the tutorial 1 scene.

Added scriptable objects.

Setting the text game objects true and false to switch between the different texts when the collisions happen.

Next scene.

Added an invoke of 25 seconds which will set the canvas with the button to go to the next scene to active.